routing admin routes correctly
