+++ cosc360-rsvp/client/src/features/adminManagement/AdminManagement.jsx
@@ -21,7 +21,7 @@ const AdminManagement = () => {
         async function fetchData() {
             try {
                 const [usersRes, eventsRes] = await Promise.all([
-                    fetch('/api/users', { headers: { 'x-user-id': activeUserId } }),
+                    fetch('/api/admin', { headers: { 'x-user-id': activeUserId } }),
                     fetch('/api/events'),
                 ]);
                 const usersData = await usersRes.json();
@@ -39,7 +39,7 @@ const AdminManagement = () => {
     async function handleDeleteUser(userId){
         if(!confirm("Are you sure you want to delete this user?")) return;
         try{
-            const response = await fetch(`/api/users/${userId}`, {
+            const response = await fetch(`/api/admin/${userId}`, {
                 method: 'DELETE',
                 headers: {'x-user-id': activeUserId},
             });
@@ -59,7 +59,7 @@ const AdminManagement = () => {
         const message = newRole === 'admin' ? "Promote this user to admin?" : "Demote this user to regular user?";
         if(!confirm(message)) return;
         try{
-            const response = await fetch(`/api/users/${userId}/role`,{
+            const response = await fetch(`/api/admin/${userId}/role`,{
                 method: 'PUT',
                 headers:{
                     'Content-Type': 'application/json',
