--- cosc360-rsvp/server/src/modules/repository/adminRepository.js
@@ -1,5 +1,0 @@
-import { UserSchema } from "../model/user.model.js";
-
-export async function getAllUsers() {
-    return UserSchema.find();
-}
