+++ cosc360-rsvp/server/src/modules/routes/userRoutes.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { getProfile, updateProfile, uploadPhoto, listUsers, deleteUser, updateUserRole, createUser, loginUser, updatePassword } from "../controllers/userController.js";
+import { getProfile, updateProfile, uploadPhoto, createUser, loginUser, updatePassword } from "../controllers/userController.js";
 import { uploadImage, authMiddleware } from "../middleware.js";
 
 const router = express.Router();
@@ -7,11 +7,8 @@ const router = express.Router();
 router.post("/register", createUser);
 router.post("/reset-password", updatePassword);
 router.post("/login", loginUser);
-router.get("/", authMiddleware, listUsers);
 router.get("/profile", getProfile);
 router.put("/profile", updateProfile);
 router.post("/profile/photo", uploadImage.single("profilePhoto"), uploadPhoto);
-router.delete("/:id", authMiddleware, deleteUser);
-router.put("/:id/role", authMiddleware, updateUserRole);
 
 export default router;
